nginx-otel-ecs-cdk/requirements.txt
attrs==22.2.0
aws-cdk-lib==2.23.0
cattrs==22.2.0
constructs==10.1.271
exceptiongroup==1.1.0
importlib-resources==5.12.0
jsii==1.77.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.5.0
zipp==3.15.0
